==> 2-Connecting Crimes to Nearest Subway Locations.ipynb <==
import pandas as pd
from datetime import datetime, timedelta, time
import numpy as np
import seaborn as sns
import sys


print("Python Version:", sys.version)
print("Pandas Version:", pd.__version__)
print("Numpy Version:", np.__version__)
print("Seaborn Version:", sns.__version__)

main_path = '/Users/PC/Desktop/Insight/MTASafety'

# spherical distance
# Taken from here: 
# https://community.esri.com/groups/coordinate-reference-systems/blog/2017/10/05/haversine-formula
def haversine(coord1: object, coord2: object):
    import math

    # Coordinates in decimal degrees (e.g. 2.89078, 12.79797)
    lon1, lat1 = coord1
    lon2, lat2 = coord2

    R = 6371000  # radius of Earth in meters
    phi_1 = math.radians(lat1)
    phi_2 = math.radians(lat2)

    delta_phi = math.radians(lat2 - lat1)
    delta_lambda = math.radians(lon2 - lon1)

    a = math.sin(delta_phi / 2.0) ** 2 + math.cos(phi_1) * math.cos(phi_2) * math.sin(delta_lambda / 2.0) ** 2
    
    c = 2 * math.atan2(math.sqrt(a), math.sqrt(1 - a))

    meters = R * c  # output distance in meters
    km = meters / 1000.0  # output distance in kilometers

    meters = round(meters, 3)
    km = round(km, 3)
    miles = round(meters/1609.34,3)
    
    return miles
    #print(f"Distance: {meters} m")
    #print(f"Distance: {km} km")

Crimes=pd.read_csv(main_path + '/CrimeData/CrimeData2.csv')
Stations=pd.read_csv(main_path + '/RiderData/TurnstileStationsMapping6.csv')

list(zip(Stations['Longitude'],Stations['Latitude']))

#Creating Long/Lat Lists
Stations['LONG/LAT']=list(zip(Stations['Longitude'],Stations['Latitude']))
Crimes['LONG/LAT']=list(zip(Crimes['Longitude'],Crimes['Latitude']))

#Calculating the Matrix Size
Crimes2 = Crimes.reset_index()
num_rows = Crimes2.shape[0]

Stations2 = Stations.reset_index()
num_cols = Stations2.shape[0]


# Creating the distance matrix
dist_mat = np.zeros([num_rows, num_cols])
for i in range(0, num_rows):
    for j in range(0, num_cols):
        dist_mat[i,j] = haversine(Crimes2.loc[i, 'LONG/LAT'], Stations2.loc[j, 'LONG/LAT'])

#Distance DataFrame Column/Row Names
ComplaintNum=list(Crimes['CMPLNT_NUM'])
StationName=list(Stations['Station2'])

#Creating the Distance Dataframe
Distance=pd.DataFrame(dist_mat, index=ComplaintNum, columns=StationName)

#Getting the lowest distance Station
CrimesMatch = Distance.idxmin(axis=1)

CrimesMatch.head()

#Getting the distance of the lowest distance Station
CrimesDist = Distance.min(axis=1)

CrimesDist.head()

Crimes['Nearest Station']=CrimesMatch.values

Crimes['Distance']=CrimesDist.values

Crimes.head()

Crimes.to_csv(main_path + '/CrimeData/CrimeData6.csv')

Distance.to_csv(main_path + '/CrimeData/DistanceMatrix2.csv')

#Based on just geolocation how many crimes happen "outside" of a 1/8th of a mile from a subway station
Crimes.query('Distance > .125').shape
#Handcoding the errors
